Refactor providers & routes
