+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "./shared/App";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
