+++ src/pages/Login.jsx
@@ -2,11 +2,14 @@ import React, { useContext } from "react";
 import { Redirect } from "react-router-dom";
 import firebase from "firebase";
 import { FirebaseAuth } from "react-firebaseui";
+
 import { AuthContext } from "../context/auth";
 
 function Login(props) {
   const { user } = useContext(AuthContext);
-  const referer = { ...props }.location.state.referer || "/";
+  const referer = { ...props }.location.state
+    ? { ...props }.location.state.referer
+    : "/";
 
   const uiConfig = {
     signInFlow: "popup",
